coffin/requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
blinker==1.4
braceexpand==0.1.7
cffi==1.16.0
click==8.1.7
cryptography==3.4.8
dbus-python==1.2.18
discord-ios==1.1.1
disdick @ git+https://github.com/topfloorchris/disdick@1f456132718b6f7f9a83b73ab27e10062ccaaba5
distro==1.7.0
durations-nlp==1.0.1
fail2ban==0.11.2
frozenlist==1.4.1
httplib2==0.20.2
humanize==4.9.0
idna==3.6
import-expression==1.1.4
importlib-metadata==4.6.4
jeepney==0.7.1
jishaku==2.5.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
multidict==6.0.5
munch==4.0.0
netifaces==0.11.0
oauthlib==3.2.0
orjson==3.9.15
pomice==2.9.0
psutil==5.9.8
pycares==4.4.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyYAML==5.4.1
rembg==2.0.61
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
typing_extensions==4.10.0
ufw==0.36.1
wadllib==1.3.6
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
zipp==1.0.0